update sources codes 11.12

+++ src/main/scala/EdgeWeightUpdate.scala
@@ -2,8 +2,8 @@ import org.apache.spark.graphx._
 
 object EdgeWeightUpdate {
   def updateEdgeWeight(
-    edgeWeightUpdateGraph: Graph[(Long, Long), Double],
-    oldEdgeWeights: Array[Double]): Array[Double] ={
+    edgeWeightUpdateGraph: Graph[(Long, Long), Long],
+    oldEdgeWeights: Array[Double]): Array[Double] = {
 
     val newEdgeWeights = Array(1.0, 1.2, 0.8, 1.0)
